pyproject.toml
[project]
name = "matensemble"
version = "0.1.4"
description = "An adaptive and highly asynchronous ensemble simulation workflow manager MatEnsemble (https://github.com/Q-CAD/MatEnsemble) built jointly on top of the hierarchical graph based scheduler FLUX and concurrent-futures infrastructure of python"
readme = "README.md"
license-files = ["LICENSE"]
authors = [
    { name = "Soumendu Bagchi", email = "soumendubagchi@gmail.com" },
    { name = "Kaleb Duchesneau", email = "kalebduchesneau@gmail.com" },
]
requires-python = ">=3.12"
dependencies = [
    "ase>=3.22.0",
    "matplotlib>=3.4.0",
    "mpi4py>=3.1.0",
    "networkx>=3.6.1",
    "numpy>=1.21.0",
    "ovito>=3.7.0",
    "pandas>=1.3.0",
    "pymatgen>=2022.0.0",
    "redis>=7.2.0",
    "scikit-learn>=1.0.0",
    "scipy>=1.7.0",
    "seaborn>=0.11.0",
]

[project.optional-dependencies]
flux = [
    "flux-python==0.66.0",
]

[build-system]
requires = ["uv_build>=0.10.0,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "myst-parser>=5.0.0",
    "pytest>=9.0.2",
    "sphinx>=9.1.0",
    "sphinx-autodoc-typehints>=3.6.2",
    "sphinx-rtd-theme>=3.1.0",
]
